Create README.md
Each script is titled and can be matched with the correct .csv files for use in the analysis via the table provided in the document titled "Data Frames and R Scripts for Statistical Analyses".   The table provides the names of the statistical tests conducted, the site and season from which the data used in the analysis was collected, the titles of the .csv files (data frames) that contain the data, and the title of the R script containing the code for performing the statistical analysis.  The    .csv files are available in the Supplementary Information accompanying the publication "Improved Total Mercury Analysis in High C and S Estuarine Porewaters using Longer Oxidation Times".